requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
Django==3.1.1
django-crispy-forms==1.11.2
django-jsonfield==1.4.1
idna==2.10
isort==5.5.1
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
python-docx==0.8.10
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
webdriver-manager==3.2.2
wrapt==1.12.1
